+++ domain/User.h
@@ -8,12 +8,12 @@ class User
         User(const std::string& username, const std::string& displayName, const std::string& password);
         std::string getUsername() const;
         std::string getDisplayName() const;
-        uint8_t getUserID() const;
+        uint16_t getUserID() const;
         bool checkPassword(const std::string& password) const;
 
     private:
         std::string m_username;
         std::string m_displayName;
-        uint8_t m_userID;
+        uint16_t m_userID;
         std::string m_password;
 };
